Resolved: Cropped landscape image shows areas that are not included in calculations #7
Updated all plots to display the buffered landscape, with the exception of the final top combination plot which does not display the buffer (the logic being that no combinations can occur in the buffer). This can be changed to also display the buffer if preferred. 

Fixed a bug where the default 1200 metre buffer would not trigger unless the input box was interacted with. Interaction is not required for the 1200 metre default now.